Added login redirect
The user will now be redirected to /u/<username>/notes when logged in.
