+++ src/main/java/de/thowl/notetilus/web/RegisterController.java
@@ -1,0 +1,35 @@
+package de.thowl.notetilus.web;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import de.thowl.notetilus.core.services.AuthenticationService;
+import de.thowl.notetilus.web.forms.RegisterForm;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Controller
+public class RegisterController {
+
+	@Autowired
+	private AuthenticationService authsvc;
+
+	@GetMapping("/signup")
+	public String showRegiterPage() {
+		log.info("entering showRegisterPage (GET-Method: /register)");
+		return "signup";
+	}
+
+	@PostMapping("/signup")
+	public String doRegister(RegisterForm form, Model model) {
+		log.info("entering doRegister (POST-Method: /register)");
+	
+		authsvc.register(form.getUsername(), form.getEmail(), form.getPassword(), form.getPassword2());
+
+		return "index";
+	}
+}
