+++ src/main/resources/data.sql
@@ -8,4 +8,4 @@ VALUES
 INSERT INTO
     users(id, username, email, password, group_id)
 VALUES
-    (0, 'Rüdiger', 'ruediger@thowl.de', '$2a$15$q/VCgMfG.XLG55DT2dgehOVFcAMgoV2iwgB7AA2yhzpk53E6uJ.La', 1);
+    (0, 'Ruediger', 'ruediger@thowl.de', '$2a$15$q/VCgMfG.XLG55DT2dgehOVFcAMgoV2iwgB7AA2yhzpk53E6uJ.La', 1);
