+++ src/main/resources/templates/login.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="de_DE" xmlns:th="https://www.thymeleaf.org">
   <head>
-    <title>BCrypt Demo</title>
+    <title>Notetilus</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
