+++ src/main/resources/templates/notes.html
@@ -1,0 +1,22 @@
+<!DOCTYPE HTML>
+<html lang="de_DE" xmlns:th="https://www.thymeleaf.org">
+  <head>
+    <title>Notilus</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
+    rel="stylesheet"
+    integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
+    crossorigin="anonymous">
+  </head>
+  <body>
+    <nav class="navbar navbar-light bg-light">
+      <div class="container-fluid">
+        <img th:src="@{images/thowl_logo.png}" alt="Logo der th-owl" width="100"/>
+      </div>
+    </nav>
+    
+    This is the notes page, if you see this you are logged in.
+
+  </body>
+ </html>
